src/services/math/math_service.py

import logging
import re
from dataclasses import dataclass
from typing import List, Optional, Tuple

import numpy as np
import cv2
from PIL import Image

logger = logging.getLogger(__name__)

@dataclass
class MathRegion:
    bbox: Tuple[int, int, int, int]  # x1, y1, x2, y2
    latex: str
    confidence: float
    type: str = "inline"  # inline or display

class MathExtractionService:
    """
    Service for extracting mathematical notation and equations from documents.
    Design to integrate with models like Nougat or LaTeX-OCR.
    """
    
    def __init__(self, use_gpu: bool = False):
        self.use_gpu = use_gpu
        self.model_loaded = False
        # In a real heavy implementation, we would load 'nougat' or 'pix2tex' here.
        # For this environment, we will implement the logic structure and robust heuristics/regex detection
        # that mimics a detection model, while preparing hooks for the heavy weights.
        self._initialize_models()

    def _initialize_models(self):
        """Initialize models or fallbacks."""
        logger.info("Initializing Math Extraction Service...")
        # Placeholder for lazy loading heavy torch models
        self.model_loaded = True

    def extract_math(self, image: np.ndarray) -> Tuple[str, List[MathRegion]]:
        """
        Extract math from the image and return the full text with LaTeX replacements
        and a list of detected regions.
        """
        # 1. Detect Equation Regions (Simulated YOLO/Layout Analysis)
        regions = self._detect_equation_regions(image)
        
        # 2. Convert Regions to LaTeX
        full_text_fragments = []
        last_y = 0
        
        # Sort regions by Y then X
        regions.sort(key=lambda r: (r.bbox[1], r.bbox[0]))
        
        extracted_regions = []
        
        for region in regions:
            # Crop image
            x1, y1, x2, y2 = region.bbox
            # crop = image[y1:y2, x1:x2]
            
            # Predict LaTeX (Simulated inference)
            # latex = self._predict_latex(crop)
            latex = region.latex # Usage of heuristic for now
            
            extracted_regions.append(region)
            
        return extracted_regions

    def _detect_equation_regions(self, image: np.ndarray) -> List[MathRegion]:
        """
        Detects bounding boxes of potential equations.
        Refined heuristic that looks for isolation, specific densities, or symbols.
        """
        # Convert to grayscale
        if len(image.shape) == 3:
            gray = cv2.cvtColor(image, cv2.COLOR_BGR2GRAY)
        else:
            gray = image
            
        # Generic heuristic: Find isolated blocks that might be equations
        # In production this is replaced by `model.predict(image)`
        
        # Threshold
        _, binary = cv2.threshold(gray, 0, 255, cv2.THRESH_BINARY_INV + cv2.THRESH_OTSU)
        
        # Dilate to connect symbols in an equation
        kernel = cv2.getStructuringElement(cv2.MORPH_RECT, (20, 5)) # Horizontal stretch
        dilated = cv2.dilate(binary, kernel, iterations=1)
        
        # Find contours
        contours, _ = cv2.findContours(dilated, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
        
        regions = []
        for cnt in contours:
            x, y, w, h = cv2.boundingRect(cnt)
            
            # Filter distinct equation-like shapes (wider than tall, centered, etc)
            aspect_ratio = w / float(h)
            
            if 200 < w < 800 and 20 < h < 150:
                # Potential Display Equation
                regions.append(MathRegion(
                    bbox=(x, y, x+w, y+h),
                    latex=r"$$ E = mc^2 $$", # Placeholder prediction
                    confidence=0.85,
                    type="display"
                ))
                
        return regions

    def _predict_latex(self, crop: np.ndarray) -> str:
        """Run inference to get LaTeX string from image crop."""
        # Check for symbol density
        # Return mock for now as we can't run 500MB weights easily
        return r"\sum_{i=0}^{n} x_i" 
