pyproject.toml
[project]
name = "document-ingestion-service"
version = "1.0.0"
description = "Intelligent Document Ingestion Service with OCR, Classification, and Extraction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    # Web Framework
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
    "python-multipart>=0.0.6",
    "jinja2>=3.1.2",
    "aiofiles>=23.2.1",
    
    # Async Task Queue
    "celery[redis]>=5.3.0",
    "redis>=5.0.0",
    
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "asyncpg>=0.29.0",
    "alembic>=1.12.0",
    
    # OCR & Image Processing
    "paddlepaddle>=2.5.0",
    "paddleocr>=2.7.0",
    "opencv-python>=4.8.0",
    "pillow>=10.0.0",
    "numpy>=1.24.0",
    
    # PDF Processing
    "pypdf>=3.17.0",
    "pdf2image>=1.16.0",
    
    # LLM Integration
    "openai>=1.3.0",
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    
    # Data Validation
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    
    # Testing & QA
    "jiwer>=3.0.0",
    "augraphy>=8.2.0",
    
    # Dashboard
    "streamlit>=1.28.0",
    "plotly>=5.18.0",
    
    # Load Testing
    "locust>=2.19.0",
    
    # Utilities
    "python-dotenv>=1.0.0",
    "httpx>=0.25.0",
    "aiofiles>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "black>=23.11.0",
    "ruff>=0.1.6",
    "mypy>=1.7.0",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
